move config pkg
