+++ backend/cmd/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	_ "github.com/lib/pq"
-	"github.com/yoshihiro-shu/draft-backend/config"
 	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server"
+	"github.com/yoshihiro-shu/draft-backend/internal/config"
 )
 
 func main() {
